+++ tests/tasks/databricks/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest.importorskip("pydantic")
